refactor(avatar): rename of component multi word name
