+++ packages/frameworks/vue/src/avatar/AvatarRoot.vue
